Fix crypto libraries for 32-bit

+++ src/crypto/md_helper.c
@@ -51,7 +51,7 @@
  * ==========================(LICENSE BEGIN)============================
  *
  * Copyright (c) 2007-2010  Projet RNRT SAPHIR
- *
+ * 
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -59,10 +59,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * 
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * 
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
@@ -75,7 +75,6 @@
  *
  * @author   Thomas Pornin <thomas.pornin@cryptolog.com>
  */
-
 
 #ifdef _MSC_VER
 #pragma warning (disable: 4146)
@@ -130,7 +129,7 @@ SPH_XCAT(sph_, HASH)(void *cc, const void *data, size_t len)
 #endif
 {
 	SPH_XCAT(sph_, SPH_XCAT(HASH, _context)) *sc;
-	unsigned current;
+	size_t current;
 
 	sc = cc;
 #if SPH_64
@@ -139,7 +138,7 @@ SPH_XCAT(sph_, HASH)(void *cc, const void *data, size_t len)
 	current = (unsigned)sc->count_low & (SPH_BLEN - 1U);
 #endif
 	while (len > 0) {
-		unsigned clen;
+		size_t clen;
 #if !SPH_64
 		sph_u32 clow, clow2;
 #endif
